refactor switchboardLabel -> switch
