+++ identity/app/views/profile/privacyForm.scala.html
@@ -15,7 +15,7 @@
   availableLists: EmailNewsletters)(implicit request: RequestHeader, messages: play.api.i18n.Messages)
 
 @consentCheckbox(consentField: Field) = {
-    @fragments.form.switchboardLabel(
+    @fragments.form.switch(
         title = getConsentText(consentField).toString(),
         subheading = None,
         description = None,
@@ -52,7 +52,7 @@
         <h2 class="manage-account-header">Would you like to receive information from the Guardian and their partners?</h2>
         <p class="manage-account-headerNote">The Guardian and their partners would like to occasionally send you information about their products, services and events</p>
         <div class="fieldset fieldset--manage-account-noborder fieldset--manage-account-block">
-            <div class="manage-account__switchboard">
+            <div class="manage-account__switches">
                 <ul>
                 @helper.repeat(privacyForm("consents"), min=1) { consentField =>
                     <li>
