+++ static/src/javascripts/projects/common/modules/identity/modules/switch.js
